return undefined in extractSubCommands
